+++ NEWS.md
@@ -1,3 +1,7 @@
 # aniSNA 1.1
 
-* Added a `NEWS.md` file to track changes to the package.
+## Functions made flexible to be used with any choice of network metrics
+
+aniSNA 1.1.1
+
+##Function names modified to better represent their usage